complete until crud messege

+++ studybud/base/views.py
@@ -1,25 +1,23 @@
 from django.shortcuts import render,redirect
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate , login ,logout
-
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
-# from django.http import HttpResponse
-
-
-# rooms =[
-#     {'id':1 , 'name':'lets learn python'},
-#     {'id':2 , 'name':'lets learn js'},
-#     {'id':3 , 'name':'lets learn php'},
-#     {'id': 4 , 'name' : 'what the heelll'},
-# ]
+from django.http import HttpResponse
 
 def loginpage(request):
+    page = 'login'
+    
+    
+    if request.user.is_authenticated :
+        return redirect('home')
     
     if request.method=='POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password= request.POST.get('password')
         try :
             user=User.objects.get(username=username)
@@ -33,7 +31,7 @@ def loginpage(request):
         else :
             messages.error(request , 'username or Password is not correct')
             
-    context={}
+    context={'page':page}
     return render(request, 'base/login_rig.html', context)
 
 
@@ -41,6 +39,22 @@ def logoutuser(request):
     
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+    if request.method== 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user= form.save(commit =False)
+            user.username = user.username.lower()
+            user.save()
+            login(request , user)
+            return redirect('home')
+        else :
+            messages.error(request , 'An error happened during registration')   
+            
+    return render(request ,'base/login_rig.html', {'form': form} )
 
 def home(request):
     
@@ -68,6 +82,10 @@ def room(request,pk):
     
     return render(request,'base/room.html', context)
 # Create your views here.
+
+
+
+@login_required(login_url= 'login')
 def CreateRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -79,9 +97,15 @@ def CreateRoom(request):
     return render(request, 'base/room_form.html', context)
 
 
+
+@login_required(login_url= 'login')
 def updateRoom(request , pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+    if request.user!= room.host:
+        return HttpResponse('you are not allowed')
+        
+    
     if request.method =='POST':
         form = RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -90,8 +114,14 @@ def updateRoom(request , pk):
     context = {'form':form}
     return render(request , 'base/room_form.html', context)
 
+
+@login_required(login_url= 'login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+    if request.user!= room.host and not request.user.is_superuser :
+        return HttpResponse('you are not allowed')
+    
+    
     if request.method=='POST':
         room.delete()
         return redirect('home')
